refactor(scrollprogress): move to components folder
